+++ CHANGELOG.md
@@ -3,10 +3,15 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased][]
+### Features
 - **Changes**: --inverse option to show common commits [#118][] [#134][]
+
+### Changes
+- **Changes**: rename --remote to --upstream [#140][]
 
 [#118]: https://github.com/Brickster/git-commands/issues/118
 [#134]: https://github.com/Brickster/git-commands/issues/134
+[#140]: https://github.com/Brickster/git-commands/issues/140
 
 ## [v0.7.0][] - 2019-08-26
 ### Features